ALGAS/requirements.txt
mysql-connector-python==8.0.28
protobuf==3.19.4
pysimplegui==4.57.0